3296: AC.cpp (#1439) - AC,29.22%,19.01%

+++ Codes/3296-minimum-number-of-seconds-to-make-mountain-height-zero.cpp
@@ -12,7 +12,7 @@ typedef long long ll;
 class Solution {
 public:
     ll minNumberOfSeconds(int mountainHeight, vector<int>& workerTimes) {
-        priority_queue<tuple<ll, int, int>, vector<tuple<ll, int, int>>, greater<>> pq;
+        priority_queue<tuple<ll, ll, int>, vector<tuple<ll, ll, int>>, greater<>> pq;
         for (int t : workerTimes) {
             pq.push({t, 1, t});
         }
